+++ examples/query_optional.rs
@@ -57,7 +57,7 @@ fn queries(txn: &'static mut Transaction<Schema>) {
     // most powerful pattern, can retrieve optional data in one query
     let _info = txn.query_one(optional(|row| {
         let player = row.and(Player.pub_id(pub_id));
-        row.then(PlayerInfo::from_expr(player))
+        row.then_select(PlayerInfo::from_expr(player))
     }));
 
     // for simple queries, use the trivial mapping